Fixed [ 1011111 ] mishandling of UID W.R.T layers. The reference engine is now using _getCopy() mark the copy of an object with a marker _v_is_cp so manage_afterAdd() can detect copies in order to reassign a new UUID and remove the reference annotations.
From the bug report and in order to justify this larger chance inside a release candidate:
BaseObject.manage_afterAdd() calls
Referenceable.manage_afterAdd() and
self.initializeLayers(), however the same doesn't
happen on manage_afterClone() and
Referenceable.manage_afterAdd() doesn't change the UID
of objects on a copy/paset or clone situation.

The result is that any layer operation that relies on
external resources keyed by UID will sufer hard to
debug errors (picture a file system storage layer
looking up a directory by UID that it supposes to be
created on storage.initializeInstance()).

I believe the correct way to deal with this is to deal
with duplicate UIDs on Referenceable.manage_afterAdd()
instead of .manage_afterClone().

OFS.CopySupport gives us other options as well, such as
flagging the object on obj._notifyOfCopyTo() so that it
can change UID on .manage_afterAdd() or even
overloading obj._getCopy() to return a copy with a
different UID (this would be my choice).

The fix is based on the idea of Rochael, made by Ben and applied by me after I had some time to understand why it's required and how it's working.